+++ http/controller/admin/my/tag.go
@@ -1,4 +1,4 @@
-package admin
+package my
 
 import (
 	"Gwen/global"
@@ -7,42 +7,46 @@ import (
 	"Gwen/service"
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
-	"strconv"
 )
 
-type Tag struct {
-}
+type Tag struct{}
 
-// Detail 标签
-// @Tags 标签
-// @Summary 标签详情
-// @Description 标签详情
+// List 列表
+// @Tags 我的标签
+// @Summary 标签列表
+// @Description 标签列表
 // @Accept  json
 // @Produce  json
-// @Param id path int true "ID"
-// @Success 200 {object} response.Response{data=model.Tag}
+// @Param page query int false "页码"
+// @Param page_size query int false "页大小"
+// @Param is_my query int false "是否是我的"
+// @Param user_id query int false "用户id"
+// @Success 200 {object} response.Response{data=model.TagList}
 // @Failure 500 {object} response.Response
-// @Router /admin/tag/detail/{id} [get]
+// @Router /admin/my/tag/list [get]
 // @Security token
-func (ct *Tag) Detail(c *gin.Context) {
-	id := c.Param("id")
-	iid, _ := strconv.Atoi(id)
-	t := service.AllService.TagService.InfoById(uint(iid))
+func (ct *Tag) List(c *gin.Context) {
+	query := &admin.TagQuery{}
+	if err := c.ShouldBindQuery(query); err != nil {
+		response.Fail(c, 101, response.TranslateMsg(c, "ParamsError")+err.Error())
+		return
+	}
 	u := service.AllService.UserService.CurUser(c)
-	if !service.AllService.UserService.IsAdmin(u) && t.UserId != u.Id {
-		response.Fail(c, 101, response.TranslateMsg(c, "NoAccess"))
-		return
-	}
-	if t.Id > 0 {
-		response.Success(c, t)
-		return
-	}
-	response.Fail(c, 101, response.TranslateMsg(c, "ItemNotFound"))
-	return
+	query.UserId = int(u.Id)
+	res := service.AllService.TagService.List(query.Page, query.PageSize, func(tx *gorm.DB) {
+		tx.Preload("Collection", func(txc *gorm.DB) *gorm.DB {
+			return txc.Select("id,name")
+		})
+		tx.Where("user_id = ?", query.UserId)
+		if query.CollectionId != nil && *query.CollectionId >= 0 {
+			tx.Where("collection_id = ?", query.CollectionId)
+		}
+	})
+	response.Success(c, res)
 }
 
 // Create 创建标签
-// @Tags 标签
+// @Tags 我的标签
 // @Summary 创建标签
 // @Description 创建标签
 // @Accept  json
@@ -50,7 +54,7 @@ func (ct *Tag) Detail(c *gin.Context) {
 // @Param body body admin.TagForm true "标签信息"
 // @Success 200 {object} response.Response{data=model.Tag}
 // @Failure 500 {object} response.Response
-// @Router /admin/tag/create [post]
+// @Router /admin/my/tag/create [post]
 // @Security token
 func (ct *Tag) Create(c *gin.Context) {
 	f := &admin.TagForm{}
@@ -64,10 +68,8 @@ func (ct *Tag) Create(c *gin.Context) {
 		return
 	}
 	t := f.ToTag()
-	if t.UserId == 0 {
-		response.Fail(c, 101, response.TranslateMsg(c, "ParamsError"))
-		return
-	}
+	u := service.AllService.UserService.CurUser(c)
+	t.UserId = u.Id
 	err := service.AllService.TagService.Create(t)
 	if err != nil {
 		response.Fail(c, 101, response.TranslateMsg(c, "OperationFailed")+err.Error())
@@ -76,42 +78,8 @@ func (ct *Tag) Create(c *gin.Context) {
 	response.Success(c, nil)
 }
 
-// List 列表
-// @Tags 标签
-// @Summary 标签列表
-// @Description 标签列表
-// @Accept  json
-// @Produce  json
-// @Param page query int false "页码"
-// @Param page_size query int false "页大小"
-// @Param is_my query int false "是否是我的"
-// @Param user_id query int false "用户id"
-// @Success 200 {object} response.Response{data=model.TagList}
-// @Failure 500 {object} response.Response
-// @Router /admin/tag/list [get]
-// @Security token
-func (ct *Tag) List(c *gin.Context) {
-	query := &admin.TagQuery{}
-	if err := c.ShouldBindQuery(query); err != nil {
-		response.Fail(c, 101, response.TranslateMsg(c, "ParamsError")+err.Error())
-		return
-	}
-	res := service.AllService.TagService.List(query.Page, query.PageSize, func(tx *gorm.DB) {
-		tx.Preload("Collection", func(txc *gorm.DB) *gorm.DB {
-			return txc.Select("id,name")
-		})
-		if query.UserId > 0 {
-			tx.Where("user_id = ?", query.UserId)
-		}
-		if query.CollectionId != nil && *query.CollectionId >= 0 {
-			tx.Where("collection_id = ?", query.CollectionId)
-		}
-	})
-	response.Success(c, res)
-}
-
 // Update 编辑
-// @Tags 标签
+// @Tags 我的标签
 // @Summary 标签编辑
 // @Description 标签编辑
 // @Accept  json
@@ -119,7 +87,7 @@ func (ct *Tag) List(c *gin.Context) {
 // @Param body body admin.TagForm true "标签信息"
 // @Success 200 {object} response.Response{data=model.Tag}
 // @Failure 500 {object} response.Response
-// @Router /admin/tag/update [post]
+// @Router /admin/my/tag/update [post]
 // @Security token
 func (ct *Tag) Update(c *gin.Context) {
 	f := &admin.TagForm{}
@@ -136,12 +104,27 @@ func (ct *Tag) Update(c *gin.Context) {
 		response.Fail(c, 101, response.TranslateMsg(c, "ParamsError"))
 		return
 	}
+
+	u := service.AllService.UserService.CurUser(c)
+	if f.UserId != u.Id {
+		response.Fail(c, 101, response.TranslateMsg(c, "NoAccess"))
+		return
+	}
 	ex := service.AllService.TagService.InfoById(f.Id)
 	if ex.Id == 0 {
 		response.Fail(c, 101, response.TranslateMsg(c, "ItemNotFound"))
 		return
 	}
+	if ex.UserId != u.Id {
+		response.Fail(c, 101, response.TranslateMsg(c, "NoAccess"))
+		return
+	}
+
 	t := f.ToTag()
+	if t.CollectionId > 0 && !service.AllService.AddressBookService.CheckCollectionOwner(t.UserId, t.CollectionId) {
+		response.Fail(c, 101, response.TranslateMsg(c, "ParamsError"))
+		return
+	}
 	err := service.AllService.TagService.Update(t)
 	if err != nil {
 		response.Fail(c, 101, response.TranslateMsg(c, "OperationFailed")+err.Error())
@@ -159,7 +142,7 @@ func (ct *Tag) Update(c *gin.Context) {
 // @Param body body admin.TagForm true "标签信息"
 // @Success 200 {object} response.Response
 // @Failure 500 {object} response.Response
-// @Router /admin/tag/delete [post]
+// @Router /admin/my/tag/delete [post]
 // @Security token
 func (ct *Tag) Delete(c *gin.Context) {
 	f := &admin.TagForm{}
@@ -178,10 +161,16 @@ func (ct *Tag) Delete(c *gin.Context) {
 		response.Fail(c, 101, response.TranslateMsg(c, "ItemNotFound"))
 		return
 	}
+	u := service.AllService.UserService.CurUser(c)
+	if ex.UserId != u.Id {
+		response.Fail(c, 101, response.TranslateMsg(c, "NoAccess"))
+		return
+	}
 	err := service.AllService.TagService.Delete(ex)
 	if err == nil {
 		response.Success(c, nil)
 		return
 	}
 	response.Fail(c, 101, err.Error())
+	return
 }
